+++ xtask/src/main.rs
@@ -117,6 +117,10 @@ impl CommandMiri {
             "tests-integration",
             &["--test", "unsafe_paths_test"],
         ));
+        run_command(make_miri_cmd(
+            "tests-integration",
+            &["--test", "waitset_reentrancy_test"],
+        ));
     }
 }